Started with google-oauth-token-authorization, node.js mailer and options for friends and likes as well

+++ codeial/controllers/api/v1/posts_api.js
@@ -23,24 +23,24 @@ module.exports.index = async function (req, res) {
   }
 };
 
-module.exports.destroy = async function(req, body){
-  try{
+module.exports.destroy = async function (req, res) {
+  try {
     let post = await Post.findById(req.params.id);
-    if(post.user == req.user.id){
-      post.remove();
-      await Comment.deleteMany({post: req.params.id});
-      return res.json(200, {
-        message: "Post and associated commenst deleted successfully !!"
+    if (post && String(post.user) === req.user.id) {
+      await post.remove();
+      await Comment.deleteMany({ post: req.params.id });
+      return res.status(200).json({
+        message: "Post and associated comments deleted successfully!"
       });
     } else {
-      return res.json(401, {
+      return res.status(401).json({
         message: "You cannot delete this post!"
       });
     }
-  } catch(err){
-    console.log('*****', err);
-    return res.json(500, {
+  } catch (err) {
+    console.error('*****', err);
+    return res.status(500).json({
       message: "Internal Server Error"
     });
   }
-}
+};
